requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
pymysql>=1.1.0
cryptography>=41.0.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
